The First Draft of the LaTeX Beamer Slide Template and Intro Slides
A draft for a LaTeX slide template (based on the beamer class) has
been developed. This template fits more or less what I imagine the
slides here should look like. It uses various symbols of USTC and is
very similar to another USTC corporate-design based template I saw
somewhere else.

Also the first draft of the start of the slides I would use to introduce
the evaluator component has been designed. This draft, however, only
covers a few slides and is mainly a test of the slide template. It will
be filled with useful contents later on.

Additionally, I put some scripts which can compile the LaTeX documents
to PDF and translate an EPS file to a cropped PDF file.